requirements.txt
Django >= 4.2.2, < 4.3
djangorestframework >= 3.14.0, < 3.15
psycopg2 >= 2.9.6, < 2.10
drf-spectacular >= 0.26.2, < 0.27
openai >= 0.27.8, < 0.28
tiktoken >= 0.4.0, < 0.5